+++ front-end/src/App.tsx
@@ -1,26 +1,32 @@
 import React from "react";
-import "./App.scss";
 
 //Components
 import Button from "./components/button/button";
 import Input from "./components/fields/input/input";
 import Alert from "./components/alert/alert";
 import Checkbox from "./components/fields/checkbox/checkbox";
-import Dropdown from "./components/dropdown/dropdown";
+import DateTimePicker from "./components/datetimepicker/datetimepicker";
+import SearchInput from "./components/fields/input/searchInput";
+import Badge from "./components/badge/badge";
+import Toast from "./components/toast/toast";
+
+//Styles
+import "./App.scss";
 
 function App() {
 	return (
-		// Component  List
 		<div>
-			<div className="m-8 flex max-w-fit items-center justify-center gap-4 align-middle">
-				<div>
+			{/* // Component List */}
+			<div className="m-8 flex max-w-fit flex-wrap items-center justify-center align-middle">
+				<div className="flex-item flex justify-center">
 					<Button
+						globalCustomStyling=""
 						color="btn-primary"
 						buttonStyling="btn rounded-3xl w-28 text-base normal-case font-medium"
 						label="Log in"
 					/>
 				</div>
-				<div>
+				<div className="flex-item flex justify-center">
 					<Input
 						inputType="text"
 						formControlStyling="form-control w-full max-w-xs"
@@ -32,7 +38,7 @@ function App() {
 						bottomLeftLabel="&nbsp;"
 					/>
 				</div>
-				<div className="w-96">
+				<div className="flex-item flex w-96 justify-center">
 					<Alert
 						globalCustomStyling="flex w-full max-w-sm overflow-hidden rounded-lg bg-white shadow-md"
 						statusStyling="flex w-12 items-center justify-center"
@@ -44,7 +50,7 @@ function App() {
 						desc="Message successfully sent to Visitor via Email"
 					/>
 				</div>
-				<div>
+				<div className="flex-item flex justify-center">
 					<Checkbox
 						globalStyling="form-control"
 						labelStyling="label cursor-pointer"
@@ -53,9 +59,27 @@ function App() {
 						label="I have agreed to the terms and conditions"
 					/>
 				</div>
-				<div>
-					<Dropdown />
+				<div className="flex-item flex justify-center">
+					<DateTimePicker />
 				</div>
+				<div className="flex-item flex justify-center">
+					<SearchInput globalStyling="mx-auto w-96" placeHolder="Search" />
+				</div>
+				<div className="flex-item flex justify-center gap-2">
+					<Badge status="approved" textSize="text-base" />
+					<Badge status="in-progress" textSize="text-base" />
+					<Badge status="declined" textSize="text-base" />
+				</div>
+				<div className="flex-item flex justify-center gap-2">
+					<Badge status="approved" textSize="text-base" />
+					<Badge status="in-progress" textSize="text-base" />
+					<Badge status="declined" textSize="text-base" />
+				</div>
+			</div>
+			<div className="flex-item flex justify-center gap-2">
+				<Toast status="info" toastInfo="Information Notification" />
+				{/* <Toast status="success" toastInfo="Successful Notification" />
+					<Toast status="error" toastInfo="Notification Error" /> */}
 			</div>
 		</div>
 	);